Move avl.{h,c} to subdirectory
